Add a method to build a docker image
Sometimes it is desirable to build an image from source instead of
pulling it. This commit adds the necessary function to do that.
