+++ pkg/client/docker/daemon.go
@@ -8,7 +8,6 @@ import (
 	"encoding/json"
 	"errors"
 	"fmt"
-	"io"
 	"net"
 	"os"
 	"path/filepath"
@@ -152,23 +151,6 @@ func ConnectDaemon(ctx context.Context, address string) (conn *grpc.ClientConn, 
 		}
 		return conn, nil
 	}
-}
-
-// PullImage checks if the given image exists locally by doing docker image inspect. A docker pull is
-// performed if no local image is found. Stdout is silenced during those operations.
-func PullImage(ctx context.Context, image string) error {
-	cmd := proc.StdCommand(ctx, "docker", "image", "inspect", image)
-	cmd.Stderr = io.Discard
-	cmd.Stdout = io.Discard
-	if cmd.Run() == nil {
-		// Image exists in the local cache, so don't bother pulling it.
-		return nil
-	}
-	cmd = proc.StdCommand(ctx, "docker", "pull", image)
-	// Docker run will put the pull logs in stderr, but docker pull will put them in stdout.
-	// We discard them here, so they don't spam the user. They'll get errors through stderr if it comes to it.
-	cmd.Stdout = io.Discard
-	return cmd.Run()
 }
 
 const kubeAuthPortFile = kubeauth.CommandName + ".port"
